feat(memory): MemoryServer MCP adapter — memory.retrieve_similar_experience (MET-453)
Exposes the L2 agent-memory contract as an MCP tool the harness can
invoke. Mirrors the knowledge adapter's layering: tool_registry imports
only the published L2 contract (MemoryClient + MemorySearchHit), with
the layer exception documented in the module docstring.

* tool_registry/tools/memory/adapter.py — MemoryServer subclass of
  McpToolServer with the memory.retrieve_similar_experience tool.
  Inputs: goal (required string), limit (1..50), agent_code, only_success.
  Project scoping is read from mcp_core.current_context (MET-401/387)
  rather than the wire payload so callers can't bypass tenant isolation.
* tests/unit/test_memory_mcp_adapter.py — 8 unit tests covering manifest
  registration, lazy-binding error, top-hit ordering, agent/success
  filters, validation errors, and wire-shape serialization (UUIDs +
  datetimes coerced to strings).

Tag-alongs:
* digital_twin/memory/store.py — clamp similarity to [-1, 1] to absorb
  cosine float drift so MemorySearchHit's range validator never trips.
* digital_twin/memory/importance.py — explicit float() coercion on the
  recency decay return to satisfy mypy --strict no_any_return.

Quality gates: ruff clean, mypy --strict clean (9 source files),
36 tests passing (28 prior + 8 new).

+++ digital_twin/memory/importance.py
@@ -99,7 +99,7 @@ def _recency_score(event: Event, now: datetime, half_life_hours: float) -> float
     age_hours = age_seconds / 3600.0
     if half_life_hours <= 0:
         return 1.0 if age_hours == 0 else 0.0
-    return 0.5 ** (age_hours / half_life_hours)
+    return float(0.5 ** (age_hours / half_life_hours))
 
 
 def _relevance_score(event: Event) -> float:
